removed the method to remove an archived course. added new method to update pastGpa added annotations
