+++ src/main/java/accountClasses/AccountInteractor.java
@@ -1,4 +1,4 @@
-package AccountClasses;
+package accountClasses;
 
 import java.util.*;
